requirements.txt
Flask==3.1.0
stomp.py==8.2.0